feat： 添加用户表存储用户关系

+++ src/main/java/org/liahnu/bot/model/domain/Contest.java
@@ -1,13 +1,13 @@
 package org.liahnu.bot.model.domain;
 
 import com.baomidou.mybatisplus.annotation.*;
+import lombok.Data;
+import org.liahnu.bot.model.type.ContestStatus;
+import org.liahnu.bot.model.type.ContestType;
 
 import java.io.Serial;
 import java.io.Serializable;
 import java.util.Date;
-import lombok.Data;
-import org.liahnu.bot.model.type.ContestStatus;
-import org.liahnu.bot.model.type.ContestType;
 
 /**
  * 比赛
@@ -38,7 +38,7 @@ public class Contest implements Serializable {
      * 
      */
     @TableField(value = "create_user_id")
-    private Long createUserId;
+    private Integer createUserId;
 
     /**
      *
